feat(docker): improve Docker setup and seeding process
- Remove volume cleanup from docker-clean-up script to avoid data loss.
- Prevent seeding in production by adding an environment check in seeds.rb.
- Persist database data in Docker by adding a volume to docker-compose.yml.

+++ db/seeds.rb
@@ -9,6 +9,12 @@
 #   end
 #
 # # Clear existing data (optional, if you want to start fresh)
+
+if Rails.env.production?
+  puts "Seeding is disabled in production!"
+  exit
+end
+
 Article.destroy_all
 
 article1 = Article.create!(
